Refactoring structure to be more modulus
to create folder structure was using "real life java 9" example
