+++ Java/MineSweeper/game.mineSweeper.core/src/game/mineSweeper/core/GameMineMap.java
@@ -1,4 +1,4 @@
-package mineSweeper.core;
+package game.mineSweeper.core;
 
 import java.util.function.Predicate;
