+++ Java/MineSweeper/game.mineSweeper.core/src/game/mineSweeper/core/NearByPositions.java
@@ -1,4 +1,4 @@
-package mineSweeper.core;
+package game.mineSweeper.core;
 
 
 import java.util.*;
