+++ Java/MineSweeper/game.mineSweeper.core/src/module-info.java
@@ -1,0 +1,2 @@
+module game.mineSweeper.core {
+}
