+++ Java/MineSweeper/game.mineSweeper.core/test/game/mineSweeper/core/PositionTest.java
@@ -1,5 +1,6 @@
-package mineSweeper.core;
+package game.mineSweeper.core;
 
+import game.mineSweeper.core.Position;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
