+++ Java/MineSweeper/game.mineSweeper.gui/src/module-info.java
@@ -1,0 +1,2 @@
+module game.mineSweeper.gui {
+}
