+++ Java/MineSweeper/game.mineSweeper.solver/src/module-info.java
@@ -1,0 +1,2 @@
+module game.mineSweeper.solver {
+}
